add and update files: add inc folder and update index file of user

+++ users/index.php
@@ -1,13 +1,6 @@
-<?php require("../process/config.php"); ?>
-<!doctype html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-+0n0xVW2eSR5OomGNYDnhzAbDsOXxcvSN1TPprVMTNDbiYZCxYbOOl7+AMvyTG2x" crossorigin="anonymous">
-    <title>Title</title>
-</head>
+<?php include("../inc/header.php"); ?>
 <body>
+<?php include("../inc/navbar.php"); ?>
     <section>
         <div class="container py-5">
             <div class="card">
@@ -39,7 +32,7 @@
                                     <td>
                                         <a name="" id="" class="btn btn-info btn-sm" href="edit.php?id=<?php echo $data['id'];?>" role="button">Edit</a>
                                         <a name="" id="" class="btn btn-primary btn-sm" href="view.php?id=<?php echo $data['id'];?>" role="button">View</a>
-                                        <a name="" id="" class="btn btn-danger btn-sm" href="../process/delete-user.php?id=<?php echo $data['id'];?>" role="button">Delete</a>
+                                        <a name="" id="" class="btn btn-danger btn-sm" href="../process/delete-user.php?id=<?php echo $data['id'];?>" onclick="return confirm('Are you to delete data??');" role="button">Delete</a>
                                     </td>
                                 </tr>
                             <?php
@@ -51,6 +44,4 @@
             </div>
         </div>
     </section>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-gtEjrD/SeCtmISkJkNUaaKMoLD0//ElJ19smozuHV6z3Iehds+3Ulb9Bn9Plx0x4" crossorigin="anonymous"></script>
-</body>
-</html>
+    <?php include("../inc/footer.php"); ?>
